Finish replaceable event store extraction
